Move snapshot manual tests
